+++ src/configure-templates/patch-template/1.copy-source-files.sh
@@ -4,6 +4,6 @@ source "#OPAL_TOOLS_USER_ENV_SCRIPT#"
 SCRIPT_DIR=`dirname $0`
 #@REM Yellow forground color for Integration / T&A
 #color 0E
-opal-install-copy-source-files.sh --source-path "$OPAL_TOOLS_SRC_SQL_DIR" \
-                                  --target-path "${SCRIPT_DIR}/sql" \
-                                  --patch-file-name "${SCRIPT_DIR}/SourceFilesCopy.txt"
+opal-install-copy-source-files.sh --source-dir "$OPAL_TOOLS_SRC_SQL_DIR" \
+                                  --target-dir "${SCRIPT_DIR}/sql" \
+                                  --source-list-file "${SCRIPT_DIR}/SourceFilesCopy.txt"